Move writer out of the checker into another file.
